pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
